Avoid duplicate cskiplist type declarations by reading a new header.
