+++ src/lib/prof-lean/cskiplist.h
@@ -87,29 +87,7 @@
 typedef struct cskiplist_s cskiplist_t;
 
 #else
-
-#include "mcs-lock.h"
-#include "pfq-rwlock.h"
-
-typedef struct csklnode_s {
-  void *val;
-  int height;
-  volatile bool fully_linked;
-  volatile bool marked;
-  mcs_lock_t lock;
-  // memory allocated for a node will include space for its vector of  pointers
-  struct csklnode_s *nexts[];
-} csklnode_t;
-
-typedef struct cskiplist_s {
-  csklnode_t *left_sentinel;
-  csklnode_t *right_sentinel;
-  int max_height;
-  val_cmp compare;
-  val_cmp inrange;
-  pfq_rwlock_t lock;
-} cskiplist_t;
-
+#include "cskiplist_defs.h"
 #endif
 
 /*
